finish 29 homework

+++ src/feature/blogger/services/blogger-service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateBlogInputModel } from '../../blogs/api/pipes/create-blog-input-model';
 import { CreateBlog } from '../../blogs/api/types/dto';
 import { BlogSqlTypeormRepository } from '../../blogs/repositories/blog-sql-typeorm-repository';
@@ -31,14 +31,42 @@ export class BloggerService {
     return this.blogSqlTypeormRepository.createNewBlogForCorrectUser(newBlog);
   }
 
-  async updateBlog(blogId: string, updateBlogInputModel: CreateBlogInputModel) {
+  async updateBlog(
+    blogId: string,
+    updateBlogInputModel: CreateBlogInputModel,
+    userId: string,
+  ) {
+    const blog =
+      await this.blogSqlTypeormRepository.getBlogByBlogIdWithUserInfo(blogId);
+
+    if (!blog) {
+      return false;
+    }
+
+    if (!blog.usertyp || blog.usertyp.id !== userId) {
+      /*   403 статус код */
+      throw new ForbiddenException('forbidden to put blogs');
+    }
+
     return this.blogSqlTypeormRepository.updateBlog(
       blogId,
       updateBlogInputModel,
     );
   }
 
-  async deleteBlogById(blogId: string) {
+  async deleteBlogById(blogId: string, userId: string) {
+    const blog =
+      await this.blogSqlTypeormRepository.getBlogByBlogIdWithUserInfo(blogId);
+
+    if (!blog) {
+      return false;
+    }
+
+    if (!blog.usertyp || blog.usertyp.id !== userId) {
+      /*   403 статус код */
+      throw new ForbiddenException('forbidden to put blogs');
+    }
+
     return this.blogSqlTypeormRepository.deleteBlogById(blogId);
   }
 }
